move sequence serialization tests to kernel tests
