+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -109,24 +109,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "sequence_dataset_serialization_test",
-    size = "medium",
-    srcs = ["sequence_dataset_serialization_test.py"],
-    shard_count = 5,
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-    ],
-)
-
-tf_py_test(
     name = "serialization_integration_test",
     size = "small",
     srcs = ["serialization_integration_test.py"],
